% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.matrix.plateLayout}
\alias{as.matrix.plateLayout}
\title{Returns a matrix form of the plate layout}
\usage{
\method{as.matrix}{plateLayout}(x, ...)
}
\arguments{
\item{x}{a plateLayout class object}

\item{...}{further arguments (ignored)}
}
\value{
a matrix of the well types with dimnames
}
\description{
Returns a matrix form of the plate layout
}
\seealso{
Other LAYOUT: \code{\link{colnames.plateLayout}};
  \code{\link{countWells}};
  \code{\link{create.plateLayout}};
  \code{\link{dim.plateLayout}};
  \code{\link{dimnames.plateLayout}};
  \code{\link{getWellType}}; \code{\link{getWellnames}};
  \code{\link{is.plateLayout}}; \code{\link{plateLayout}},
  \code{\link{plateLayout}},
  \code{\link{plateLayout-package}};
  \code{\link{read.plateLayout}};
  \code{\link{rownames.plateLayout}};
  \code{\link{to_tecan}}; \code{\link{write.plateLayout}}
}

