% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{check.wellname}
\alias{check.wellname}
\title{Checks a well name to make sure it follows the A01 format}
\usage{
check.wellname(well, reverse = FALSE)
}
\arguments{
\item{well}{character vector of well names}

\item{reverse}{logical, if TRUE convert "A01"... to  "A:1" "B:2" "C:3"}
}
\value{
a character vector of well names
}
\description{
Checks a well name to make sure it follows the A01 format
}

