% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{check.wellnames}
\alias{check.wellnames}
\title{A wrapper around \code{\link{check.wellname}}}
\usage{
check.wellnames(well, ...)
}
\arguments{
\item{well}{character vector of well names}

\item{...}{further arguments for \code{\link{check.wellname}}}
}
\value{
a character vector of well names
}
\description{
A wrapper around \code{\link{check.wellname}}
}
