% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{colnames.plateLayout}
\alias{colnames.plateLayout}
\title{Returns the column names of the plateLayout object as if it were a matrix}
\usage{
colnames.plateLayout(x)
}
\arguments{
\item{x}{plateLayout class object}
}
\value{
a character vector of the column names
}
\description{
Returns the column names of the plateLayout object as if it were a matrix
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}};
  \code{\link{countWells}};
  \code{\link{create.plateLayout}};
  \code{\link{dim.plateLayout}};
  \code{\link{dimnames.plateLayout}};
  \code{\link{getWellType}}; \code{\link{getWellnames}};
  \code{\link{is.plateLayout}}; \code{\link{plateLayout}},
  \code{\link{plateLayout}},
  \code{\link{plateLayout-package}};
  \code{\link{read.plateLayout}};
  \code{\link{rownames.plateLayout}};
  \code{\link{to_tecan}}; \code{\link{write.plateLayout}}
}

