% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{countWells}
\alias{countWells}
\title{Returns the number of wells in a layout}
\usage{
countWells(x)
}
\arguments{
\item{x}{a plateLayout}
}
\value{
numeric, number of wells
}
\description{
Returns the number of wells in a layout
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}};
  \code{\link{colnames.plateLayout}};
  \code{\link{create.plateLayout}};
  \code{\link{dim.plateLayout}};
  \code{\link{dimnames.plateLayout}};
  \code{\link{getWellType}}; \code{\link{getWellnames}};
  \code{\link{is.plateLayout}}; \code{\link{plateLayout}},
  \code{\link{plateLayout}},
  \code{\link{plateLayout-package}};
  \code{\link{read.plateLayout}};
  \code{\link{rownames.plateLayout}};
  \code{\link{to_tecan}}; \code{\link{write.plateLayout}}
}

