% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{create.plateLayout}
\alias{create.plateLayout}
\title{Creates a plateLayout with the names attached and type represented by 
the a random selection of \code{ntype}.}
\usage{
create.plateLayout(nwell = 384, ntype = 3, to.matrix = FALSE)
}
\arguments{
\item{nwell}{integer, the number of wells on the plate}

\item{ntype}{integer, the number of unique dummy types to sample from LETTERS}

\item{to.matrix}{logical, if TRUE return a matrix}
}
\description{
Creates a plateLayout with the names attached and type represented by 
the a random selection of \code{ntype}.
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}},
  \code{\link{colnames.plateLayout}},
  \code{\link{countWells}},
  \code{\link{dimnames.plateLayout}},
  \code{\link{getWellType}}, \code{\link{getWellnames}},
  \code{\link{is.plateLayout}}, \code{\link{plateLayout}},
  \code{\link{read.plateLayout}},
  \code{\link{rownames.plateLayout}},
  \code{\link{to_tecan}}, \code{\link{write.plateLayout}}
}
