% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate_Layout.R
\name{Create_Plate_Layout}
\alias{Create_Plate_Layout}
\title{Create and instance of Plate_Layout with fake data}
\usage{
Create_Plate_Layout(nwell = 384, ntype = 3)
}
\arguments{
\item{nwell}{integer, the number of wells on the plate}

\item{ntype}{integer, the number of unique dummy types to sample from LETTERS}
}
\value{
an instance of Plate_Layout reference class
}
\description{
Create and instance of Plate_Layout with fake data
}
\seealso{
Other PLATE_LAYOUT: \code{\link{Plate_Layout_check_names}},
  \code{\link{Plate_Layout_colnames}},
  \code{\link{Plate_Layout_count}},
  \code{\link{Plate_Layout_dimnames}},
  \code{\link{Plate_Layout_dim}},
  \code{\link{Plate_Layout_get_names}},
  \code{\link{Plate_Layout_get_types}},
  \code{\link{Plate_Layout_read_file}},
  \code{\link{Plate_Layout_read_text}},
  \code{\link{Plate_Layout_rownames}},
  \code{\link{Plate_Layout_show}},
  \code{\link{Plate_Layout_to_tecan}},
  \code{\link{Plate_Layout_write_file}},
  \code{\link{Plate_Layout}}
}
