% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{dim.plateLayout}
\alias{dim.plateLayout}
\title{Returns the dimensions of a nwell plateLayout}
\usage{
\method{dim}{plateLayout}(x)
}
\description{
@family LAYOUT 
 @export
 @method dim plateLayout
 @param x either a plateLayout or an integer
 @return a two element vector of [nrows, ncols]
}

