% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{getWellnames}
\alias{getWellnames}
\title{Returns a vector of names A01, A02,A03,...}
\usage{
getWellnames(x, to.matrix = FALSE)
}
\arguments{
\item{x}{a plateLayout object or an integer (384, 96,  12 etc)}

\item{to.matrix}{logical, if TRUE then the returned value is an appropriately sized
matrix rather than a vector}
}
\value{
a vector (or matrix) of the wellnames
}
\description{
Returns a vector of names A01, A02,A03,...
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}},
  \code{\link{colnames.plateLayout}},
  \code{\link{countWells}},
  \code{\link{create.plateLayout}},
  \code{\link{dimnames.plateLayout}},
  \code{\link{getWellType}}, \code{\link{is.plateLayout}},
  \code{\link{plateLayout}},
  \code{\link{read.plateLayout}},
  \code{\link{rownames.plateLayout}},
  \code{\link{to_tecan}}, \code{\link{write.plateLayout}}
}
