% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{getWellType}
\alias{getWellType}
\title{Get a welltype for given well(s)}
\usage{
getWellType(x, well, check = TRUE)
}
\arguments{
\item{x}{a plateLayout object}

\item{well}{character vector of one or more well names}

\item{check}{logical, check well name for A01 format?}
}
\value{
returns a named vector, one element for each input well name
}
\description{
Get a welltype for given well(s)
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}};
  \code{\link{colnames.plateLayout}};
  \code{\link{countWells}};
  \code{\link{create.plateLayout}};
  \code{\link{dim.plateLayout}};
  \code{\link{dimnames.plateLayout}};
  \code{\link{getWellnames}}; \code{\link{is.plateLayout}};
  \code{\link{plateLayout}}, \code{\link{plateLayout}},
  \code{\link{plateLayout-package}};
  \code{\link{read.plateLayout}};
  \code{\link{rownames.plateLayout}};
  \code{\link{to_tecan}}; \code{\link{write.plateLayout}}
}

