% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_A01}
\alias{is_A01}
\title{Determine is well name is in A01 format}
\usage{
is_A01(x = c("A01", "A:1", "P:24"))
}
\arguments{
\item{x}{string, one or more well names}
}
\value{
logical, TRUE if the input is in A01 format
}
\description{
This is just guessing by looking for the A:N format.  It can be fooled.
}
