% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{loadMetaForm}
\alias{loadMetaForm}
\title{Fix metadata template}
\usage{
loadMetaForm(xlsFile)
}
\arguments{
\item{xlsFile}{The metadata template Excel file}
}
\value{
The cleaned metadata as a tibble and exports a tsv version
}
\description{
Load and transform our CRUK CI Bioinfomratics Core metadata template into a
table suitable for the PlateLayout package
}
\details{
The function takes the path for an Excel file containing completed 
CRUK CI metadata template and transforms it into a table suitable for use
with the \code{\link{randomizeSinglePlate}} function.

When loading the Excel file the function skips the first row, 
making the headers from the second (hidden) row in the template. The third
row in the template contains a contents description for each column and so
is removed from the table.

Any columns that are completely empty are removed. \code{NA} is
replaced with a blank characer vector.
}
\examples{
metadataFile <- system.file("extdata", "metadata_template_example.xls",
                           package = "PlateLayout")
metadataTable <- loadMetaForm(metadataFile)
metadataTable
}
