% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plate_Layout.R
\name{Plate_Layout}
\alias{Plate_Layout}
\title{Create an instance of a Plate_Layout from text, plateLayout object or file}
\usage{
Plate_Layout(x, asText = FALSE)
}
\arguments{
\item{x}{either a character vector, plateLayout class, OR a filename}

\item{asText}{logical indicating the input is a character vector to parse}
}
\value{
an instance of Plate_Layout reference class
}
\description{
Create an instance of a Plate_Layout from text, plateLayout object or file
}
\seealso{
Other PLATE_LAYOUT: \code{\link{Create_Plate_Layout}};
  \code{\link{Plate_Layout_check_names}};
  \code{\link{Plate_Layout_colnames}};
  \code{\link{Plate_Layout_count}};
  \code{\link{Plate_Layout_dimnames}};
  \code{\link{Plate_Layout_dim}};
  \code{\link{Plate_Layout_get_names}};
  \code{\link{Plate_Layout_get_types}};
  \code{\link{Plate_Layout_read_file}};
  \code{\link{Plate_Layout_read_text}};
  \code{\link{Plate_Layout_rownames}};
  \code{\link{Plate_Layout_show}};
  \code{\link{Plate_Layout_to_tecan}};
  \code{\link{Plate_Layout_write_file}}
}

