% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plate_Layout.R
\name{Plate_Layout_dimnames}
\alias{Plate_Layout_dimnames}
\title{Returns the dimnames of the plate as if it were a matrix}
\value{
the dim names of wells
}
\description{
Returns the dimnames of the plate as if it were a matrix
}
\seealso{
Other PLATE_LAYOUT: \code{\link{Create_Plate_Layout}};
  \code{\link{Plate_Layout_check_names}};
  \code{\link{Plate_Layout_colnames}};
  \code{\link{Plate_Layout_count}};
  \code{\link{Plate_Layout_dim}};
  \code{\link{Plate_Layout_get_names}};
  \code{\link{Plate_Layout_get_types}};
  \code{\link{Plate_Layout_read_file}};
  \code{\link{Plate_Layout_read_text}};
  \code{\link{Plate_Layout_rownames}};
  \code{\link{Plate_Layout_show}};
  \code{\link{Plate_Layout_to_tecan}};
  \code{\link{Plate_Layout_write_file}};
  \code{\link{Plate_Layout}}
}

