% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate_Layout.R
\name{Plate_Layout_get_names}
\alias{Plate_Layout_get_names}
\title{Get the names of the wells}
\arguments{
\item{n}{numeric number of wells (optional).  If not provided then the length of the 
current definition is used}

\item{asMatrix}{logical, if TRUE return the names in a matrix}
}
\value{
either a matrix or vector of well names or NULL
}
\description{
Get the names of the wells
}
\seealso{
Other PLATE_LAYOUT: \code{\link{Create_Plate_Layout}},
  \code{\link{Plate_Layout_check_names}},
  \code{\link{Plate_Layout_colnames}},
  \code{\link{Plate_Layout_count}},
  \code{\link{Plate_Layout_dimnames}},
  \code{\link{Plate_Layout_dim}},
  \code{\link{Plate_Layout_get_types}},
  \code{\link{Plate_Layout_read_file}},
  \code{\link{Plate_Layout_read_text}},
  \code{\link{Plate_Layout_rownames}},
  \code{\link{Plate_Layout_show}},
  \code{\link{Plate_Layout_to_tecan}},
  \code{\link{Plate_Layout_write_file}},
  \code{\link{Plate_Layout}}
}

