% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plate_Layout.R
\name{Plate_Layout_get_types}
\alias{Plate_Layout_get_types}
\title{Get one or more well types}
\arguments{
\item{well}{character vector of one or more well names, by default all are provided}

\item{check}{logical, test if well name is in A01 format?}

\item{asMatrix}{logical, if TRUE return the names in a matrix}
}
\value{
named character vector
}
\description{
Get one or more well types
}
\seealso{
Other PLATE_LAYOUT: \code{\link{Create_Plate_Layout}};
  \code{\link{Plate_Layout_check_names}};
  \code{\link{Plate_Layout_colnames}};
  \code{\link{Plate_Layout_count}};
  \code{\link{Plate_Layout_dimnames}};
  \code{\link{Plate_Layout_dim}};
  \code{\link{Plate_Layout_get_names}};
  \code{\link{Plate_Layout_read_file}};
  \code{\link{Plate_Layout_read_text}};
  \code{\link{Plate_Layout_rownames}};
  \code{\link{Plate_Layout_show}};
  \code{\link{Plate_Layout_to_tecan}};
  \code{\link{Plate_Layout_write_file}};
  \code{\link{Plate_Layout}}
}

