% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plate_Layout.R
\name{Plate_Layout_read_text}
\alias{Plate_Layout_read_text}
\title{Read layout information from a character vector}
\arguments{
\item{txt}{the input charcater vector, list, data.frame or matrix.  Wellnames for
the latter two are guessed at.  If a vector or list then the names provided are used.}
}
\value{
the length of the data vector invisibly
}
\description{
Read layout information from a character vector
}
\seealso{
Other PLATE_LAYOUT: \code{\link{Create_Plate_Layout}};
  \code{\link{Plate_Layout_check_names}};
  \code{\link{Plate_Layout_colnames}};
  \code{\link{Plate_Layout_count}};
  \code{\link{Plate_Layout_dimnames}};
  \code{\link{Plate_Layout_dim}};
  \code{\link{Plate_Layout_get_names}};
  \code{\link{Plate_Layout_get_types}};
  \code{\link{Plate_Layout_read_file}};
  \code{\link{Plate_Layout_rownames}};
  \code{\link{Plate_Layout_show}};
  \code{\link{Plate_Layout_to_tecan}};
  \code{\link{Plate_Layout_write_file}};
  \code{\link{Plate_Layout}}
}

