% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate_Layout.R
\name{Plate_Layout_to_tecan}
\alias{Plate_Layout_to_tecan}
\title{Translates A01, A02 names to Tecan position assignments}
\value{
a character vector of Tecan position invisibly
}
\description{
Tecan assigns position number for each in the first column, then by each in 
the second column so that A01 -> 1, B01 -> 2, C01 -> 3 and so on
}
\seealso{
Other PLATE_LAYOUT: \code{\link{Create_Plate_Layout}},
  \code{\link{Plate_Layout_check_names}},
  \code{\link{Plate_Layout_colnames}},
  \code{\link{Plate_Layout_count}},
  \code{\link{Plate_Layout_dimnames}},
  \code{\link{Plate_Layout_dim}},
  \code{\link{Plate_Layout_get_names}},
  \code{\link{Plate_Layout_get_types}},
  \code{\link{Plate_Layout_read_file}},
  \code{\link{Plate_Layout_read_text}},
  \code{\link{Plate_Layout_rownames}},
  \code{\link{Plate_Layout_show}},
  \code{\link{Plate_Layout_write_file}},
  \code{\link{Plate_Layout}}
}

