% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plateLayout-package.r, R/plateLayout.R
\docType{package}
\name{plateLayout}
\alias{plateLayout}
\alias{plateLayout-package}
\title{plateLayout}
\usage{
plateLayout(x)
}
\arguments{
\item{x}{a data.frame, list, vector or matrix of plate layout descriptions}
}
\value{
a plateLayout class vector
}
\description{
Plate layout defintions are used to identify well types and locations
  on 384 or 96 well plates. This package provides convenience functionality for
  reading writing and querying layout definitions.

A plateLayout is a character vector class.  The value of the vector is
the character description of the well type (i.e. a label) and the names
of each element are the names of the well placement.  We use a A01 format
description instead of A:1.  Wells are stored in A01, A02, A03 order along the
length of the vector.
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}};
  \code{\link{colnames.plateLayout}};
  \code{\link{countWells}};
  \code{\link{create.plateLayout}};
  \code{\link{dim.plateLayout}};
  \code{\link{dimnames.plateLayout}};
  \code{\link{getWellType}}; \code{\link{getWellnames}};
  \code{\link{is.plateLayout}};
  \code{\link{read.plateLayout}};
  \code{\link{rownames.plateLayout}};
  \code{\link{to_tecan}}; \code{\link{write.plateLayout}}
}

