% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{plotPlate}
\alias{plotPlate}
\title{Plot Plate Layout}
\usage{
plotPlate(dat, plotCol, wellCols = NULL)
}
\arguments{
\item{dat}{The plate layout tables}

\item{plotCol}{The column of the table by which to colour the wells}

\item{wellCols}{A named vector of columns for colouring the wells. Names
should match the sample groups in plotCol}
}
\value{
A ggplot object
}
\description{
Plot the plate with wells coloured according to a column in the samplesheet
if there is column called Replicate the numbers will also be printed
}
\examples{
library(readr)
designSheet <- system.file("extdata", "metadata_12x3.tsv",
                           package = "PlateLayout")
bColumns <- c("ExtractionInformation", "PassageNumber") 
designTable <- read_tsv(designSheet)
# Create an R object in the session
plateLayout <- randomizeSinglePlate(designTable,  
                                    batchColumns = bColumns,
                                    nIter = 100) 
plotPlate(plateLayout, "SampleGroup")
}
