% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{randomizeSinglePlate}
\alias{randomizeSinglePlate}
\title{Randomize Single Plate}
\usage{
randomizeSinglePlate(designSheet, outputFile,
  primaryGroup = "SampleGroup", batchColumns = NULL, nIter = 10000,
  nCores = 4)
}
\arguments{
\item{designSheet}{The path to a tab delimited sample metadata file.}

\item{outputFile}{The root name for the various output files.}

\item{primaryGroup}{The name of a column in the metadata table to use for
optimising the distribution of samples across the plate}

\item{batchColumns}{A character vector wit the names of additional columns
in the metadata table to use in assessing the distribution of the samples
across the plate.}

\item{nIter}{The number of random layouts to generate and assess.}

\item{nCores}{The number of cores to use. Set to 1 to run serial.}
}
\value{
Nothing. The function writes out the plate layout to a tsv file and
a plot for each of `primaryGroup` and `batchColumns` to png files.
}
\description{
Randomize up to 96 samples across a single 96 well plate.
}
\details{
The sample metadata file should be a tab delimited table. The 
column names can be anything. By default the function expects one column 
named \code{SampleGroup} that it will use to distribute the samples across
the plate. This can be set to any other column using the \code{primaryGroup}
argument.

The columns specified in \code{batchColumns} are used in addition
to \code{primaryGroup} to assess which of the \code{nIter} layouts is 
optimal in terms of the distribution of samples. This assessed based on the
sum of Euclidean distances between replicate samples within each column.
}
\examples{
designSheet <- system.file("extdata", "metadata_template_TEST_12x3.tsv",
                           package = "PlateLayout")
outputFile <- "Test"
bColumns <- c("ExtractionInformation", "PassageNumber") 
randomizeSinglePlate(designSheet, outputFile, batchColumns = bColumns,
                     nIter = 100) 
}
