% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.plateLayout}
\alias{read.plateLayout}
\title{Read a tab delimited plateLayout file}
\usage{
read.plateLayout(file = NULL)
}
\arguments{
\item{file}{character, the fully qualified filename}
}
\value{
a plateLayout class object
}
\description{
Layout files define for well plates the "type" for each well.  Generally, these
files are created in a spreadsheet and saved in tab-delimited form for subsequent
use.
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}};
  \code{\link{colnames.plateLayout}};
  \code{\link{countWells}};
  \code{\link{create.plateLayout}};
  \code{\link{dim.plateLayout}};
  \code{\link{dimnames.plateLayout}};
  \code{\link{getWellType}}; \code{\link{getWellnames}};
  \code{\link{is.plateLayout}}; \code{\link{plateLayout}},
  \code{\link{plateLayout}},
  \code{\link{plateLayout-package}};
  \code{\link{rownames.plateLayout}};
  \code{\link{to_tecan}}; \code{\link{write.plateLayout}}
}

