% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rownames.plateLayout}
\alias{rownames.plateLayout}
\title{Returns the row names of the plateLayout object as if it were a matrix}
\usage{
rownames.plateLayout(x)
}
\arguments{
\item{x}{plateLayout class object}
}
\value{
a character vector of the row names
}
\description{
Returns the row names of the plateLayout object as if it were a matrix
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}};
  \code{\link{colnames.plateLayout}};
  \code{\link{countWells}};
  \code{\link{create.plateLayout}};
  \code{\link{dim.plateLayout}};
  \code{\link{dimnames.plateLayout}};
  \code{\link{getWellType}}; \code{\link{getWellnames}};
  \code{\link{is.plateLayout}}; \code{\link{plateLayout}},
  \code{\link{plateLayout}},
  \code{\link{plateLayout-package}};
  \code{\link{read.plateLayout}}; \code{\link{to_tecan}};
  \code{\link{write.plateLayout}}
}

