% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate_Layout.R
\name{tecan_A01}
\alias{tecan_A01}
\title{Create a lookup table for translating tecan (1,2,3,...) to A01 (A01, B01, C01,...)}
\usage{
tecan_A01(nwell = 384)
}
\arguments{
\item{nwell}{numeric the number of wells in the container}
}
\value{
named character vector A01 = 1, B01 = 2, etc)
}
\description{
Create a lookup table for translating tecan (1,2,3,...) to A01 (A01, B01, C01,...)
}

