% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{to_tecan}
\alias{to_tecan}
\title{Translates A01, A02 names to Tecan position assignments}
\usage{
to_tecan(x)
}
\arguments{
\item{x}{a plateLayout class object}
}
\value{
a numeric vector of Tecan position
}
\description{
Tecan assigns position number for each in the first column, then by each in
the second column so that A01 -> 1, B01 -> 2, C01 -> 3 and so on
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}};
  \code{\link{colnames.plateLayout}};
  \code{\link{countWells}};
  \code{\link{create.plateLayout}};
  \code{\link{dim.plateLayout}};
  \code{\link{dimnames.plateLayout}};
  \code{\link{getWellType}}; \code{\link{getWellnames}};
  \code{\link{is.plateLayout}}; \code{\link{plateLayout}},
  \code{\link{plateLayout}},
  \code{\link{plateLayout-package}};
  \code{\link{read.plateLayout}};
  \code{\link{rownames.plateLayout}};
  \code{\link{write.plateLayout}}
}

