% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{write.plateLayout}
\alias{write.plateLayout}
\title{Write a tab-delimited plateLayout file}
\usage{
write.plateLayout(x, file = "plateLayout.txt")
}
\arguments{
\item{x}{a plateLayout class object}

\item{file}{the fully qualified output filename}
}
\description{
Write a tab-delimited plateLayout file
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}};
  \code{\link{colnames.plateLayout}};
  \code{\link{countWells}};
  \code{\link{create.plateLayout}};
  \code{\link{dim.plateLayout}};
  \code{\link{dimnames.plateLayout}};
  \code{\link{getWellType}}; \code{\link{getWellnames}};
  \code{\link{is.plateLayout}}; \code{\link{plateLayout}},
  \code{\link{plateLayout}},
  \code{\link{plateLayout-package}};
  \code{\link{read.plateLayout}};
  \code{\link{rownames.plateLayout}};
  \code{\link{to_tecan}}
}

