% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{dim.plateLayout}
\alias{dim.plateLayout}
\title{Returns the dimensions of a nwell plateLayout}
\usage{
\method{dim}{plateLayout}(x)
}
\arguments{
\item{x}{either a plateLayout or an integer}
}
\value{
a two element vector of [nrows, ncols]
}
\description{
Returns the dimensions of a nwell plateLayout
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}};
  \code{\link{colnames.plateLayout}};
  \code{\link{countWells}};
  \code{\link{create.plateLayout}};
  \code{\link{dimnames.plateLayout}};
  \code{\link{getWellType}}; \code{\link{getWellnames}};
  \code{\link{is.plateLayout}}; \code{\link{plateLayout}},
  \code{\link{plateLayout}},
  \code{\link{plateLayout-package}};
  \code{\link{read.plateLayout}};
  \code{\link{rownames.plateLayout}};
  \code{\link{to_tecan}}; \code{\link{write.plateLayout}}
}

