% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateLayout.R
\name{is.plateLayout}
\alias{is.plateLayout}
\title{Test is the object is of class plateLayout}
\usage{
is.plateLayout(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
logical, TRUE if the object is of class plateLayout
}
\description{
Test is the object is of class plateLayout
}
\seealso{
Other LAYOUT: \code{\link{as.matrix.plateLayout}},
  \code{\link{colnames.plateLayout}},
  \code{\link{countWells}},
  \code{\link{create.plateLayout}},
  \code{\link{dimnames.plateLayout}},
  \code{\link{getWellType}}, \code{\link{getWellnames}},
  \code{\link{plateLayout}},
  \code{\link{read.plateLayout}},
  \code{\link{rownames.plateLayout}},
  \code{\link{to_tecan}}, \code{\link{write.plateLayout}}
}

